[BUILD] Reverted ptxas change and fixed bug in cache key computation (#1971)
